Modify library glm3
Do not own the name of a shader program. Store only a reference to a
GLuint passed in the constructor.

This is necessary since it is simply not possible for this class to
known in advance when the shader program is no longer needed.

Remove call to glDeleteProgram() in destructor.